Add dynamic view form for user profile
