+++ src/ploneintranet/userprofile/browser/templates/userprofile.pt
@@ -113,61 +113,29 @@
 
                         </tal:block>
                         <tal:block condition="not:view/is_me">
-                            <dl class="pat-grid-list">
+                            <dl class="pat-grid-list"
+				tal:define="fields view/fields_for_display">
                                 <dt class="icon-user" i18n:translate="">
                                     Name
                                 </dt>
                                 <dd>
-                                    <em class="title" tal:condition="context/person_title|nothing" tal:content="context/person_title" /> <span tal:replace="context/fullname" />
-                                </dd>
-
-                                <dt class="icon-mail" i18n:translate="">
-                                    E-mail
-                                </dt>
+				    <em class="title" tal:condition="context/person_title|nothing" tal:content="context/person_title" /> <span tal:replace="context/fullname" />
+				</dd>
+				<tal:field repeat="field fields">
+                                <dt class="${field/icon_class}" i18n:translate=""
+				    tal:content="field/label" />
                                 <dd>
-                                    <a i18n:attributes="title" tal:attributes="title string:Mail ${context/fullname} via ${context/email}; href string:mailto:${context/email}" tal:content="context/email" />
+				    <span tal:replace="structure field/html" />
                                 </dd>
-
-                                <dt class="icon-phone" i18n:translate="">
-                                    Phone number
-                                </dt>
-                                <dd>
-                                    <a i18n:attributes="title" tal:attributes="title string:Call ${context/fullname} at ${context/telephone}; href string:tel:${context/telephone}" tal:content="context/telephone"></a>
-                                </dd>
-
-                                <dt class="icon-phone" i18n:translate="">
-                                    Mobile phone number
-                                </dt>
-                                <dd>
-                                    <a i18n:attributes="title" tal:attributes="title string:Call ${context/fullname} at ${context/mobile}; href string:tel:${context/mobile}" tal:content="context/mobile"></a>
-                                </dd>
-
-                                <dt class="icon-clock" i18n:translate="">
-                                    Time zone
-                                </dt>
-                                <dd tal:content="context/time_zone">
-                                </dd>
-
-                                <dt class="icon-globe" i18n:translate="">
-                                    Primary location
-                                </dt>
-                                <dd tal:define="primary_location view/primary_location">
-                                    <a target="new" tal:attributes="href string:https://www.google.com/maps/place/${primary_location}" tal:content="primary_location" />
-                                </dd>
-
-                                <dt class="icon-right-open" i18n:translate="">
-                                    Job title
-                                </dt>
-                                <dd tal:content="context/job_title" />
-
-                                <dt class="icon-building" i18n:translate="">
-                                    Department
-                                </dt>
-                                <dd tal:content="context/department" />
+				</tal:field>
                             </dl>
+			    <article class="pat-rich"
+				     tal:condition="context/biography">
+				<p tal:content="context/biography">
+				    Some content goes here on Jeff
+				</p>
+			    </article>
                         </tal:block>
-
-                            <article class="pat-rich" tal:content="context/biography" />
 
                     </article>
 		    <div id="person-followers" tal:define="users followers">
